LocationCard: FIX fail to load placeholder image at build time
Fixes bug where Gatsby would crash when trying to add placeholder image
for location that did not have a 2D image.

Fix included making separate components for standard thumbnail and
placeholder thumbnail, instead of switching which image to load.

Also adds new required package "gatsby-source-filesystem", which is a
required plugin to ensure that the placeholder image file can be loaded
at build time.

Note that the error messages was quite vague and it not very specific
about what exactly is causing the error, and also reports other info
that is unrelated and somewhat misleading.

Changes originally created by Colin Carver.